+++ graal/com.oracle.graal.compiler/src/com/oracle/graal/compiler/GraalDebugInitializationParticipant.java
@@ -26,6 +26,7 @@ import com.oracle.graal.debug.Debug;
 import com.oracle.graal.debug.Debug.Params;
 import com.oracle.graal.debug.DebugInitializationParticipant;
 import com.oracle.graal.debug.GraalDebugConfig;
+import com.oracle.graal.debug.TTY;
 import com.oracle.graal.serviceprovider.ServiceProvider;
 
 /**
@@ -43,17 +44,47 @@ public class GraalDebugInitializationParticipant implements DebugInitializationP
         if ("".equals(GraalDebugConfig.Options.Count.getValue())) {
             params.enableUnscopedCounters = true;
         }
+        if ("".equals(GraalDebugConfig.Options.MethodMeter.getValue())) {
+            params.enableUnscopedMethodMetrics = true;
+            // mm requires full debugging support
+            params.enable = true;
+        }
         if ("".equals(GraalDebugConfig.Options.Time.getValue())) {
             params.enableUnscopedTimers = true;
         }
         if ("".equals(GraalDebugConfig.Options.TrackMemUse.getValue())) {
             params.enableUnscopedMemUseTrackers = true;
         }
-        // unscoped counters/timers should respect method filter semantics
-        if (!params.enable && (params.enableUnscopedMemUseTrackers || params.enableUnscopedCounters || params.enableUnscopedTimers) &&
-                        GraalDebugConfig.Options.MethodFilter.getValue() != null) {
+        // unscoped metrics/timers should respect method filter semantics
+        if (!params.enable && (params.enableUnscopedMemUseTrackers || params.enableUnscopedMethodMetrics || params.enableUnscopedCounters || params.enableUnscopedTimers) &&
+                        GraalDebugConfig.isNotEmpty(GraalDebugConfig.Options.MethodFilter)) {
             params.enable = true;
-            params.enableMethodFilter = true;
+            params.enabledMethodFilter = true;
+        }
+
+        if (!params.enableUnscopedMethodMetrics && GraalDebugConfig.Options.MethodMeter.getValue() != null) {
+            // mm requires full debugging support
+            params.enable = true;
+        }
+
+        if (GraalDebugConfig.isMethodMetricsDebugValueInterceptionEnabled()) {
+            if (!params.enable) {
+                TTY.println("WARNING: MethodMeter is disabled but MethodMeterInterceptDebugValues is enabled. Ignoring MethodMeter and MethodMeterInterceptDebugValues.");
+            } else {
+                parseMethodMetricsDebugValueInterception(params);
+            }
+        }
+    }
+
+    private static void parseMethodMetricsDebugValueInterception(Params params) {
+        String[] flags = GraalDebugConfig.Options.MethodMeterInterceptDebugValues.getValue().split(":");
+        if (flags.length == 3) {
+            params.interceptMeter = flags[0].equals("X");
+            params.interceptTime = flags[1].equals("X");
+            params.interceptMem = flags[2].equals("X");
+        } else {
+            TTY.println("WARNING: Ignoring MethodMeterInterceptDebugValues as the supplied argument does not conform to the format X|O:X|O:X|O.");
+            GraalDebugConfig.Options.MethodMeterInterceptDebugValues.setValue(null);
         }
     }
 }