+++ graal/com.oracle.graal.hotspot/src/com/oracle/graal/hotspot/CompilationTask.java
@@ -30,6 +30,20 @@ import static com.oracle.graal.compiler.GraalCompilerOptions.PrintCompilation;
 import static com.oracle.graal.compiler.GraalCompilerOptions.PrintFilter;
 import static com.oracle.graal.compiler.GraalCompilerOptions.PrintStackTraceOnException;
 import static com.oracle.graal.compiler.phases.HighTier.Options.Inline;
+
+import com.oracle.graal.code.CompilationResult;
+import com.oracle.graal.debug.Debug;
+import com.oracle.graal.debug.Debug.Scope;
+import com.oracle.graal.debug.DebugCloseable;
+import com.oracle.graal.debug.DebugCounter;
+import com.oracle.graal.debug.DebugDumpScope;
+import com.oracle.graal.debug.DebugTimer;
+import com.oracle.graal.debug.Management;
+import com.oracle.graal.debug.TTY;
+import com.oracle.graal.debug.TimeSource;
+import com.oracle.graal.options.OptionValue;
+import com.oracle.graal.options.OptionValue.OverrideScope;
+
 import jdk.vm.ci.code.BailoutException;
 import jdk.vm.ci.code.CodeCacheProvider;
 import jdk.vm.ci.code.CompilationRequestResult;
@@ -46,18 +60,6 @@ import jdk.vm.ci.hotspot.events.EventProvider.CompilationEvent;
 import jdk.vm.ci.hotspot.events.EventProvider.CompilerFailureEvent;
 import jdk.vm.ci.runtime.JVMCICompiler;
 import jdk.vm.ci.services.Services;
-
-import com.oracle.graal.code.CompilationResult;
-import com.oracle.graal.debug.Debug;
-import com.oracle.graal.debug.Debug.Scope;
-import com.oracle.graal.debug.DebugCloseable;
-import com.oracle.graal.debug.DebugDumpScope;
-import com.oracle.graal.debug.DebugCounter;
-import com.oracle.graal.debug.DebugTimer;
-import com.oracle.graal.debug.Management;
-import com.oracle.graal.debug.TTY;
-import com.oracle.graal.options.OptionValue;
-import com.oracle.graal.options.OptionValue.OverrideScope;
 
 //JaCoCo Exclude
 
@@ -150,6 +152,9 @@ public class CompilationTask {
      */
     private static final DebugCounter CompiledBytecodes = Debug.counter("CompiledBytecodes");
 
+    /**
+     * Time spent in code installation.
+     */
     public static final DebugTimer CodeInstallationTime = Debug.timer("CodeInstallation");
 
     @SuppressWarnings("try")
@@ -183,7 +188,7 @@ public class CompilationTask {
             final long start;
             final long allocatedBytesBefore;
             if (printAfterCompilation || printCompilation) {
-                start = System.currentTimeMillis();
+                start = TimeSource.getTimeNS();
                 allocatedBytesBefore = printAfterCompilation || printCompilation ? Lazy.threadMXBean.getThreadAllocatedBytes(threadId) : 0L;
             } else {
                 start = 0L;
@@ -210,15 +215,16 @@ public class CompilationTask {
                 filter.remove();
 
                 if (printAfterCompilation || printCompilation) {
-                    final long stop = System.currentTimeMillis();
+                    final long stop = TimeSource.getTimeNS();
+                    final long duration = (stop - start) / 1000000;
                     final int targetCodeSize = result != null ? result.getTargetCodeSize() : -1;
                     final long allocatedBytesAfter = Lazy.threadMXBean.getThreadAllocatedBytes(threadId);
                     final long allocatedBytes = (allocatedBytesAfter - allocatedBytesBefore) / 1024;
 
                     if (printAfterCompilation) {
-                        TTY.println(getMethodDescription() + String.format(" | %4dms %5dB %5dkB", stop - start, targetCodeSize, allocatedBytes));
+                        TTY.println(getMethodDescription() + String.format(" | %4dms %5dB %5dkB", duration, targetCodeSize, allocatedBytes));
                     } else if (printCompilation) {
-                        TTY.println(String.format("%-6d JVMCI %-70s %-45s %-50s | %4dms %5dB %5dkB", getId(), "", "", "", stop - start, targetCodeSize, allocatedBytes));
+                        TTY.println(String.format("%-6d JVMCI %-70s %-45s %-50s | %4dms %5dB %5dkB", getId(), "", "", "", duration, targetCodeSize, allocatedBytes));
                     }
                 }
             }