+++ graal/com.oracle.graal.hotspot/src/com/oracle/graal/hotspot/HotSpotGraalCompiler.java
@@ -46,6 +46,7 @@ import com.oracle.graal.debug.DebugEnvironment;
 import com.oracle.graal.debug.TTY;
 import com.oracle.graal.debug.TopLevelDebugConfig;
 import com.oracle.graal.debug.internal.DebugScope;
+import com.oracle.graal.debug.internal.method.MethodMetricsRootScopeInfo;
 import com.oracle.graal.hotspot.meta.HotSpotProviders;
 import com.oracle.graal.hotspot.phases.OnStackReplacementPhase;
 import com.oracle.graal.java.GraphBuilderPhase;
@@ -87,7 +88,8 @@ public class HotSpotGraalCompiler implements GraalJVMCICompiler {
         }
 
         CompilationTask task = new CompilationTask(jvmciRuntime, this, (HotSpotCompilationRequest) request, true, true);
-        try (DebugConfigScope dcs = Debug.setConfig(new TopLevelDebugConfig())) {
+        try (DebugConfigScope dcs = Debug.setConfig(new TopLevelDebugConfig());
+                        Debug.Scope s = Debug.methodMetricsScope("HotSpotGraalCompiler", MethodMetricsRootScopeInfo.create(request.getMethod()), true)) {
             return task.runCompilation();
         }
     }