+++ graal/com.oracle.graal.hotspot/src/com/oracle/graal/hotspot/HotSpotGraalRuntime.java
@@ -44,6 +44,8 @@ import com.oracle.graal.compiler.target.Backend;
 import com.oracle.graal.debug.Debug;
 import com.oracle.graal.debug.DebugEnvironment;
 import com.oracle.graal.debug.TTY;
+import com.oracle.graal.debug.internal.DebugValuesPrinter;
+import com.oracle.graal.debug.internal.method.MethodMetricsPrinter;
 import com.oracle.graal.graph.DefaultNodeCollectionsProvider;
 import com.oracle.graal.graph.NodeCollectionsProvider;
 import com.oracle.graal.hotspot.debug.BenchmarkCounters;
@@ -149,10 +151,30 @@ public final class HotSpotGraalRuntime implements HotSpotGraalRuntimeProvider, H
         }
 
         if (Debug.areUnconditionalCountersEnabled() || Debug.areUnconditionalTimersEnabled() || (Debug.isEnabled() && areScopedGlobalMetricsEnabled()) ||
-                        (Debug.isEnabled() && Debug.isMethodFilteringEnabled())) {
+                        (Debug.isEnabled() && areScopedGlobalMetricsEnabled()) || (Debug.isEnabled() && Debug.isMethodFilteringEnabled())) {
             // This must be created here to avoid loading the DebugValuesPrinter class
             // during shutdown() which in turn can cause a deadlock
-            debugValuesPrinter = new DebugValuesPrinter();
+            int mmPrinterType = 0;
+            mmPrinterType |= MethodMetricsPrinter.Options.MethodMeterPrintAscii.getValue() ? 1 : 0;
+            mmPrinterType |= MethodMetricsPrinter.Options.MethodMeterFile.getValue() != null ? 2 : 0;
+            switch (mmPrinterType) {
+                case 0:
+                    debugValuesPrinter = new DebugValuesPrinter();
+                    break;
+                case 1:
+                    debugValuesPrinter = new DebugValuesPrinter(new MethodMetricsPrinter.MethodMetricsASCIIPrinter(TTY.out));
+                    break;
+                case 2:
+                    debugValuesPrinter = new DebugValuesPrinter(new MethodMetricsPrinter.MethodMetricsCSVFilePrinter());
+                    break;
+                case 3:
+                    debugValuesPrinter = new DebugValuesPrinter(
+                                    new MethodMetricsPrinter.MethodMetricsCompositePrinter(new MethodMetricsPrinter.MethodMetricsCSVFilePrinter(),
+                                                    new MethodMetricsPrinter.MethodMetricsASCIIPrinter(TTY.out)));
+                    break;
+                default:
+                    break;
+            }
         }
 
         // Complete initialization of backends